test updated
Took 15 seconds

+++ tests/unit/svg.test.ts
@@ -9,7 +9,9 @@ describe("optimizeSvg", () => {
 
 	afterAll(() => {
 		// Cleanup
-		fs.unlinkSync(distPath);
+		if (fs.existsSync(distPath)) {
+			fs.unlinkSync(distPath);
+		}
 	});
 
 	it("optimizes an SVG file and writes it to the specified output file", async () => {
